apps sources merge wip
